fix bug in definition of matrix columns (#66)
Fix a bug that was revealed by adding the NCELL option. This allows non-cell contigs to
be used, and revealed a rarely encountered bug, simply by expanding the scope of testing.

In forming clonotype pictures, we previously lined things up in columns using similar
CDR3_AA sequences. This made it possible for a column to contain sequences that
differed in their V..J lengths, resulting in very rare downstream crashes. Now we use
(CDR3_AA, V..J length) to line things up in columns, where the length allows for a single
indel. This did not change results on any of the preexisting tests. We added a new test
for a case that previously crashed.