import highly experimental server/client code (#234)
This change also raises the number of duplicated crates from 2 to 9, however there is little effect on the debug
executable size. The duplicated crates can probably be gotten rid of if rust-bio is updated appropriately.

There are numerous questions about what the GUI should do, and issues, which we're kicking down the road.