Move core modules to separate enclone_ranger repo. (#484)
Specifically, move:

- enclone
- enclone_args
- enclone_core
- enclone_print
- enclone_proto
- enclone_ranger
- enclone_stuff
- enclone_vars

A few files from the moved crates have been moved into the remaining
crates.

This allows us to enforce stricter standards for the components on which
cellranger has direct dependencies, while preserving freedom for rapid
iteration in this repository.

enclone_ranger is a fork.  While changes have been made there to remove
stuff which is not required to be there, for the moment this repository
is depending on the commit from that repository which is the commit
before this one.

Remove references to obsolete tests and binaries.  Some were tests
which were testing things in the moved modules, which should be
moved into those modules.